WIP migrating PeopleServiceRemote
