+++ WordPressKit/WordPressKitTests/Mock Data/people-send-invitation-success.json
