+++ WordPressKit/WordPressKitTests/Mock Data/people-validate-invitation-failure.json
