+++ WordPressKit/WordPressKitTests/Mock Data/people-validate-invitation-success.json
