+++ WordPressKit/WordPressKitTests/Mock Data/site-followers-delete-auth-failure.json
